+++ app/src/parser.c
@@ -28,6 +28,7 @@
 #include "timeutils.h"
 #include "parser_print_candid.h"
 #include "parser_print_helper.h"
+#include "parser_print_protobuf.h"
 
 #if defined(TARGET_NANOX) || defined(TARGET_NANOS2)
 // For some reason NanoX requires this function
@@ -144,105 +145,6 @@ parser_error_t parser_getNumItems(const parser_context_t *ctx, uint8_t *num_item
     return parser_ok;
 }
 
-__Z_INLINE parser_error_t print_accountBytes(sender_t sender,
-                                             SendRequest *sendrequest,
-                                             char *outVal, uint16_t outValLen,
-                                             uint8_t pageIdx, uint8_t *pageCount) {
-    uint8_t address[32];
-    MEMZERO(address, sizeof(address));
-
-    zxerr_t err = crypto_principalToSubaccount(sender.data, sender.len,
-                                               sendrequest->from_subaccount.sub_account, 32,
-                                               address, sizeof(address));
-    if (err != zxerr_ok) {
-        return parser_unexpected_error;
-    }
-
-    char buffer[80];
-    err = print_hexstring(buffer, sizeof(buffer), (uint8_t *) address, 32);
-    if (err != zxerr_ok) {
-        return parser_unexpected_error;
-    }
-
-    pageString(outVal, outValLen, buffer, pageIdx, pageCount);
-
-    return parser_ok;
-}
-
-// 365.25 * 24*60*60 = 31557600
-#define ICP_YEAR_IN_SECONDS ((uint64_t)31557600)
-
-parser_error_t parser_printDelay(uint64_t value, char *buffer, uint16_t bufferSize) {
-    MEMZERO(buffer, bufferSize);
-    uint16_t index = 0;
-    uint64_t years = value / ICP_YEAR_IN_SECONDS;
-    if (years >= 1) {
-        index += fpuint64_to_str(buffer, bufferSize, years, 0);
-        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
-        MEMCPY(buffer + index, (char *) "y", 1);
-        index += 1;
-    }
-    value %= ICP_YEAR_IN_SECONDS;
-
-    uint64_t days = value / (uint64_t) (60 * 60 * 24);
-    if (days > 0) {
-        if (index > 0) {
-            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
-            MEMCPY(buffer + index, (char *) ", ", 2);
-            index += 2;
-        }
-        index += fpuint64_to_str(buffer + index, bufferSize - index, days, 0);
-        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
-        MEMCPY(buffer + index, (char *) "d", 1);
-        index += 1;
-    }
-    value %= (uint64_t) (60 * 60 * 24);
-
-    uint64_t hours = value / (uint64_t) (60 * 60);
-    if (hours > 0) {
-        if (index > 0) {
-            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
-            MEMCPY(buffer + index, (char *) ", ", 2);
-            index += 2;
-        }
-        index += fpuint64_to_str(buffer + index, bufferSize - index, hours, 0);
-        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
-        MEMCPY(buffer + index, (char *) "h", 1);
-        index += 1;
-    }
-    value %= (uint64_t) (60 * 60);
-
-    uint64_t minutes = value / (uint64_t) (60);
-    if (minutes > 0) {
-        if (index > 0) {
-            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
-            MEMCPY(buffer + index, (char *) ", ", 2);
-            index += 2;
-        }
-        index += fpuint64_to_str(buffer + index, bufferSize - index, minutes, 0);
-        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
-        MEMCPY(buffer + index, (char *) "m", 1);
-        index += 1;
-    }
-    value %= (uint64_t) (60);
-
-    uint64_t seconds = value;
-    if (seconds > 0) {
-        if (index > 0) {
-            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
-            MEMCPY(buffer + index, (char *) ", ", 2);
-            index += 2;
-        }
-        index += fpuint64_to_str(buffer + index, bufferSize - index, seconds, 0);
-        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
-        MEMCPY(buffer + index, (char *) "s", 1);
-        index += 1;
-    }
-
-    buffer[index] = 0;
-    return parser_ok;
-}
-
 static parser_error_t parser_getItemTransactionStateRead(const parser_context_t *ctx,
                                                   uint8_t displayIdx,
                                                   char *outKey, uint16_t outKeyLen,
@@ -269,11 +171,11 @@ static parser_error_t parser_getItemTransactionStateRead(const parser_context_t 
     }
 
     if (app_mode_expert()) {
-        state_read_t *fields = &parser_tx_obj.tx_fields.stateRead;
+        const state_read_t *fields = &parser_tx_obj.tx_fields.stateRead;
 
         if (displayIdx == 1) {
             snprintf(outKey, outKeyLen, "Sender ");
-            return print_textual(fields->sender.data, fields->sender.len, outVal, outValLen, pageIdx, pageCount);
+            return print_textual(fields->sender.data, (uint8_t) fields->sender.len, outVal, outValLen, pageIdx, pageCount);
         }
 
         displayIdx -= 2;
@@ -297,832 +199,6 @@ static parser_error_t parser_getItemTransactionStateRead(const parser_context_t 
     return parser_ok;
 }
 
-static parser_error_t parser_getItemTokenTransfer(const parser_context_t *ctx,
-                                                  uint8_t displayIdx,
-                                                  char *outKey, uint16_t outKeyLen,
-                                                  char *outVal, uint16_t outValLen,
-                                                  uint8_t pageIdx, uint8_t *pageCount) {
-    MEMZERO(outKey, outKeyLen);
-    MEMZERO(outVal, outValLen);
-    snprintf(outKey, outKeyLen, "?");
-    snprintf(outVal, outValLen, "?");
-    *pageCount = 1;
-
-    uint8_t numItems = 0;
-    CHECK_PARSER_ERR(parser_getNumItems(ctx, &numItems))
-    CHECK_APP_CANARY()
-
-    call_t *fields = &parser_tx_obj.tx_fields.call;
-
-    if (displayIdx < 0 || displayIdx >= numItems) {
-        return parser_no_data;
-    }
-
-    const bool is_stake_tx = parser_tx_obj.special_transfer_type == neuron_stake_transaction;
-    if (is_stake_tx) {
-        return parser_unexpected_error;
-    }
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Send ICP");
-        return parser_ok;
-    }
-
-    if (app_mode_expert()) {
-        if (displayIdx == 1) {
-            snprintf(outKey, outKeyLen, "Sender ");
-            return print_textual(fields->sender.data, fields->sender.len, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (displayIdx == 2) {
-            snprintf(outKey, outKeyLen, "Subaccount ");
-            snprintf(outVal, outValLen, "Not set");
-
-            if (fields->data.SendRequest.has_from_subaccount) {
-                char buffer[100];
-                zxerr_t err = print_hexstring(buffer, sizeof(buffer),
-                                              fields->data.SendRequest.from_subaccount.sub_account, 32);
-                if (err != zxerr_ok) {
-                    return parser_unexpected_error;
-                }
-                pageString(outVal, outValLen, buffer, pageIdx, pageCount);
-            }
-
-            return parser_ok;
-        }
-        displayIdx -= 2;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "From account");
-        return print_accountBytes(fields->sender, &fields->data.SendRequest,
-                                  outVal, outValLen,
-                                  pageIdx, pageCount);
-    }
-
-    if (displayIdx == 2) {
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.has_to, parser_unexpected_number_items)
-        snprintf(outKey, outKeyLen, "To account ");
-
-        char buffer[100];
-        zxerr_t err = print_hexstring(buffer, sizeof(buffer), (uint8_t *) fields->data.SendRequest.to.hash, 32);
-        if (err != zxerr_ok) {
-            return parser_unexpected_error;
-        }
-
-        pageString(outVal, outValLen, buffer, pageIdx, pageCount);
-        return parser_ok;
-    }
-
-    if (displayIdx == 3) {
-        snprintf(outKey, outKeyLen, "Payment (ICP)");
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.payment.has_receiver_gets, parser_unexpected_number_items)
-        return print_ICP(fields->data.SendRequest.payment.receiver_gets.e8s,
-                         outVal, outValLen,
-                         pageIdx, pageCount);
-    }
-
-    if (displayIdx == 4) {
-        snprintf(outKey, outKeyLen, "Maximum fee (ICP)");
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.has_max_fee, parser_unexpected_number_items)
-        return print_ICP(fields->data.SendRequest.max_fee.e8s,
-                         outVal, outValLen,
-                         pageIdx, pageCount);
-    }
-
-    if (displayIdx == 5) {
-        snprintf(outKey, outKeyLen, "Memo");
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.has_memo, parser_unexpected_number_items)
-        return print_u64(fields->data.SendRequest.memo.memo, outVal, outValLen, pageIdx, pageCount);
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemStakeNeuron(const parser_context_t *ctx,
-                                                uint8_t displayIdx,
-                                                char *outKey, uint16_t outKeyLen,
-                                                char *outVal, uint16_t outValLen,
-                                                uint8_t pageIdx, uint8_t *pageCount) {
-    MEMZERO(outKey, outKeyLen);
-    MEMZERO(outVal, outValLen);
-    snprintf(outKey, outKeyLen, "?");
-    snprintf(outVal, outValLen, "?");
-    *pageCount = 1;
-
-    uint8_t numItems = 0;
-    CHECK_PARSER_ERR(parser_getNumItems(ctx, &numItems))
-    CHECK_APP_CANARY()
-
-    call_t *fields = &parser_tx_obj.tx_fields.call;
-
-    if (displayIdx < 0 || displayIdx >= numItems) {
-        return parser_no_data;
-    }
-
-    const bool is_stake_tx = parser_tx_obj.special_transfer_type == neuron_stake_transaction;
-    if (!is_stake_tx) {
-        return parser_unexpected_error;
-    }
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Stake Neuron");
-        return parser_ok;
-    }
-
-    if (app_mode_expert()) {
-        if (displayIdx == 1) {
-            snprintf(outKey, outKeyLen, "Sender ");
-            return print_textual(fields->sender.data, fields->sender.len, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (displayIdx == 2) {
-            snprintf(outKey, outKeyLen, "Subaccount ");
-            snprintf(outVal, outValLen, "Not set");
-
-            if (fields->data.SendRequest.has_from_subaccount) {
-                char buffer[100];
-                zxerr_t err = print_hexstring(buffer, sizeof(buffer),
-                                              fields->data.SendRequest.from_subaccount.sub_account, 32);
-                if (err != zxerr_ok) {
-                    return parser_unexpected_error;
-                }
-                pageString(outVal, outValLen, buffer, pageIdx, pageCount);
-            }
-
-            return parser_ok;
-        }
-        displayIdx -= 2;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "From account");
-        return print_accountBytes(fields->sender, &fields->data.SendRequest,
-                                  outVal, outValLen,
-                                  pageIdx, pageCount);
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Payment (ICP)");
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.payment.has_receiver_gets, parser_unexpected_number_items)
-        return print_ICP(fields->data.SendRequest.payment.receiver_gets.e8s,
-                         outVal, outValLen,
-                         pageIdx, pageCount);
-    }
-
-    if (displayIdx == 3) {
-        snprintf(outKey, outKeyLen, "Maximum fee (ICP)");
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.has_max_fee, parser_unexpected_number_items)
-        return print_ICP(fields->data.SendRequest.max_fee.e8s,
-                         outVal, outValLen,
-                         pageIdx, pageCount);
-    }
-
-    if (displayIdx == 4) {
-        snprintf(outKey, outKeyLen, "Memo");
-        PARSER_ASSERT_OR_ERROR(fields->data.SendRequest.has_memo, parser_unexpected_number_items)
-        return print_u64(fields->data.SendRequest.memo.memo, outVal, outValLen, pageIdx, pageCount);
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemStartStopDissolve(uint8_t displayIdx,
-                                                      char *outKey, uint16_t outKeyLen,
-                                                      char *outVal, uint16_t outValLen,
-                                                      uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-
-        manageNeuron_e mn_type;
-        CHECK_PARSER_ERR(getManageNeuronType(&parser_tx_obj, &mn_type))
-
-        if (mn_type == Configure_StartDissolving) {
-            snprintf(outVal, outValLen, "Start Dissolve     Neuron");
-        } else {
-            snprintf(outVal, outValLen, "Stop Dissolve      Neuron");
-        }
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        } else if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        } else {
-            //Only accept neuron_id
-            return parser_unexpected_type;
-        }
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemSpawn(uint8_t displayIdx,
-                                          char *outKey, uint16_t outKeyLen,
-                                          char *outVal, uint16_t outValLen,
-                                          uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Spawn Neuron");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Controller ");
-        if (!fields->command.spawn.has_new_controller) {
-
-            snprintf(outVal, outValLen, "Self");
-            return parser_ok;
-        }
-
-        PARSER_ASSERT_OR_ERROR(fields->command.spawn.new_controller.serialized_id.size <= 29,
-                               parser_value_out_of_range)
-
-        return print_textual(fields->command.spawn.new_controller.serialized_id.bytes,
-                             29,
-                             outVal, outValLen,
-                             pageIdx, pageCount);
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemAddRemoveHotkey(uint8_t displayIdx,
-                                                    char *outKey, uint16_t outKeyLen,
-                                                    char *outVal, uint16_t outValLen,
-                                                    uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-
-        manageNeuron_e mn_type;
-        CHECK_PARSER_ERR(getManageNeuronType(&parser_tx_obj, &mn_type))
-
-        if (mn_type == Configure_AddHotKey) {
-            snprintf(outVal, outValLen, "Add Hotkey");
-        } else {
-            snprintf(outVal, outValLen, "Remove Hotkey");
-        }
-
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Principal ");
-        manageNeuron_e mn_type;
-        CHECK_PARSER_ERR(getManageNeuronType(&parser_tx_obj, &mn_type))
-
-        if (mn_type == Configure_AddHotKey) {
-            PARSER_ASSERT_OR_ERROR(fields->command.configure.operation.add_hot_key.has_new_hot_key,
-                                   parser_unexpected_number_items)
-            PARSER_ASSERT_OR_ERROR(fields->command.configure.operation.add_hot_key.new_hot_key.serialized_id.size <= 29,
-                                   parser_value_out_of_range)
-            return print_textual(fields->command.configure.operation.add_hot_key.new_hot_key.serialized_id.bytes, 29,
-                                 outVal, outValLen, pageIdx, pageCount);
-        }
-
-        PARSER_ASSERT_OR_ERROR(fields->command.configure.operation.remove_hot_key.has_hot_key_to_remove,
-                               parser_unexpected_number_items)
-        PARSER_ASSERT_OR_ERROR(
-                fields->command.configure.operation.remove_hot_key.hot_key_to_remove.serialized_id.size <= 29,
-                parser_value_out_of_range)
-
-        return print_textual(
-                fields->command.configure.operation.remove_hot_key.hot_key_to_remove.serialized_id.bytes, 29,
-                outVal, outValLen, pageIdx, pageCount);
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemDisburse(uint8_t displayIdx,
-                                             char *outKey, uint16_t outKeyLen,
-                                             char *outVal, uint16_t outValLen,
-                                             uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Disburse Neuron");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Disburse To ");
-
-        if (!fields->command.disburse.has_to_account) {
-            snprintf(outVal, outValLen, "Self");
-            return parser_ok;
-        }
-
-        PARSER_ASSERT_OR_ERROR(fields->command.disburse.to_account.hash.size == 32, parser_context_unexpected_size)
-
-        char buffer[80];
-        zxerr_t err = print_hexstring(buffer, sizeof(buffer),
-                                      (uint8_t *) fields->command.disburse.to_account.hash.bytes, 32);
-        if (err != zxerr_ok) {
-            return parser_unexpected_error;
-        }
-
-        pageString(outVal, outValLen, buffer, pageIdx, pageCount);
-        return parser_ok;
-    }
-
-    if (displayIdx == 3) {
-        snprintf(outKey, outKeyLen, "Amount (ICP)");
-
-        if (!fields->command.disburse.has_amount) {
-            snprintf(outVal, outValLen, "All");
-            return parser_ok;
-        }
-
-        return print_ICP(fields->command.disburse.amount.e8s, outVal, outValLen, pageIdx, pageCount);
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemIncreaseNeuronTimer(uint8_t displayIdx,
-                                                        char *outKey, uint16_t outKeyLen,
-                                                        char *outVal, uint16_t outValLen,
-                                                        uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Increase Dissolve  Delay");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Additional Delay");
-
-        if (fields->command.configure.operation.increase_dissolve_delay.additional_dissolve_delay_seconds == 0) {
-            snprintf(outVal, outValLen, "0s");
-            return parser_ok;
-        }
-
-        char buffer[100];
-        MEMZERO(buffer, sizeof(buffer));
-        uint64_t value = 0;
-        MEMCPY(&value,
-               &fields->command.configure.operation.increase_dissolve_delay.additional_dissolve_delay_seconds,
-               4);
-
-        CHECK_PARSER_ERR(parser_printDelay(value, buffer, sizeof(buffer)))
-        pageString(outVal, outValLen, buffer, pageIdx, pageCount);
-        return parser_ok;
-    }
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemMergeMaturity(uint8_t displayIdx,
-                                                  char *outKey, uint16_t outKeyLen,
-                                                  char *outVal, uint16_t outValLen,
-                                                  uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Stake Maturity");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Percentage");
-        char buffer[100];
-        MEMZERO(buffer, sizeof(buffer));
-        uint64_t value = 0;
-        MEMCPY(&value, &fields->command.merge_maturity.percentage_to_merge, 4);
-
-        if (value > 100) {
-            return parser_unexpected_value;
-        }
-
-        return print_u64(value, outVal, outValLen, pageIdx, pageCount);
-    }
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemJoinCommunityFund(uint8_t displayIdx,
-                                                      char *outKey, uint16_t outKeyLen,
-                                                      char *outVal, uint16_t outValLen,
-                                                      uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Join Community     Fund");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemRegisterVote(uint8_t displayIdx,
-                                                 char *outKey, uint16_t outKeyLen,
-                                                 char *outVal, uint16_t outValLen,
-                                                 uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Register Vote");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Proposal ID");
-        char buffer[100];
-        MEMZERO(buffer, sizeof(buffer));
-        uint64_t value = 0;
-        MEMCPY(&value, &fields->command.register_vote.proposal.id, 8);
-        return print_u64(value, outVal, outValLen, pageIdx, pageCount);
-    }
-
-    if (displayIdx == 3) {
-        snprintf(outKey, outKeyLen, "Vote");
-        ic_nns_governance_pb_v1_Vote v = fields->command.register_vote.vote;
-
-        if (v == 0) {
-            return parser_unexpected_value;
-        }
-
-        snprintf(outVal, outValLen, v == 1 ? "Yes" : "No");
-        return parser_ok;
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemFollow(uint8_t displayIdx,
-                                           char *outKey, uint16_t outKeyLen,
-                                           char *outVal, uint16_t outValLen,
-                                           uint8_t pageIdx, uint8_t *pageCount) {
-    *pageCount = 1;
-
-    ic_nns_governance_pb_v1_ManageNeuron *fields = &parser_tx_obj.tx_fields.call.data.ic_nns_governance_pb_v1_ManageNeuron;
-
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Follow");
-        return parser_ok;
-    }
-
-    if (displayIdx == 1) {
-        snprintf(outKey, outKeyLen, "Neuron ID");
-        PARSER_ASSERT_OR_ERROR(!(fields->has_id && (fields->which_neuron_id_or_subaccount == 12 ||
-                                                    fields->which_neuron_id_or_subaccount == 11)),
-                               parser_unexpected_number_items)
-
-        if (fields->has_id) {
-            return print_u64(fields->id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        if (fields->which_neuron_id_or_subaccount == 12) {
-            return print_u64(fields->neuron_id_or_subaccount.neuron_id.id, outVal, outValLen, pageIdx, pageCount);
-        }
-
-        //Only accept neuron_id
-        return parser_unexpected_type;
-    }
-
-    if (displayIdx == 2) {
-        snprintf(outKey, outKeyLen, "Topic");
-        ic_nns_governance_pb_v1_Topic topic = fields->command.follow.topic;
-
-        switch (topic) {
-            case ic_nns_governance_pb_v1_Topic_TOPIC_UNSPECIFIED : {
-                snprintf(outVal, outValLen, "Default");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_NEURON_MANAGEMENT : {
-                snprintf(outVal, outValLen, "Neuron Management");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_EXCHANGE_RATE : {
-                snprintf(outVal, outValLen, "Exchange Rate");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_NETWORK_ECONOMICS : {
-                snprintf(outVal, outValLen, "Network Economics");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_GOVERNANCE : {
-                snprintf(outVal, outValLen, "Governance");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_NODE_ADMIN : {
-                snprintf(outVal, outValLen, "Node Admin");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_PARTICIPANT_MANAGEMENT : {
-                snprintf(outVal, outValLen, "Participant Management");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_SUBNET_MANAGEMENT : {
-                snprintf(outVal, outValLen, "Subnet Management");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_NETWORK_CANISTER_MANAGEMENT : {
-                snprintf(outVal, outValLen, "Network Canister Management");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_KYC : {
-                snprintf(outVal, outValLen, "KYC");
-                return parser_ok;
-            }
-            case ic_nns_governance_pb_v1_Topic_TOPIC_NODE_PROVIDER_REWARDS : {
-                snprintf(outVal, outValLen, "Node Provider Rewards");
-                return parser_ok;
-            }
-            default: {
-                return parser_unexpected_type;
-            }
-        }
-    }
-
-    uint8_t new_displayIdx = displayIdx - 3;
-    pb_size_t follow_count = fields->command.follow.followees_count;
-
-    if (follow_count > 99) {
-        //check for number of chars, but the real limit is lower
-        return parser_unexpected_number_items;
-    }
-
-    if (follow_count == 0) {
-        if (new_displayIdx == 0) {
-            snprintf(outKey, outKeyLen, "Followees");
-            snprintf(outVal, outValLen, "None");
-            return parser_ok;
-        }
-
-        return parser_unexpected_number_items;
-    }
-
-    if (new_displayIdx < follow_count) {
-        uint64_t id = fields->command.follow.followees[new_displayIdx].id;
-        new_displayIdx++; //higher by 1
-        char buffer[100];
-        MEMZERO(buffer, sizeof(buffer));
-        uint16_t index = 0;
-        MEMCPY(buffer, (char *) "Followees (", 11);
-        index += 11;
-
-        uint8_t tens = new_displayIdx / 10;
-        if (tens > 0) {
-            char ten = (char) ('0' + tens);
-            MEMCPY(buffer + index, &ten, 1);
-            index++;
-        }
-
-        uint8_t ones = new_displayIdx % 10;
-        char one = (char) ('0' + ones);
-        MEMCPY(buffer + index, &one, 1);
-        index++;
-        MEMCPY(buffer + index, ")", 1);
-        snprintf(outKey, outKeyLen, "%s", buffer);
-        return print_u64(id, outVal, outValLen, pageIdx, pageCount);
-    }
-
-    return parser_no_data;
-}
-
-
-static parser_error_t parser_getItemClaimNeuron(uint8_t displayIdx,
-                                                char *outKey, uint16_t outKeyLen,
-                                                char *outVal, uint16_t outValLen) {
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "Claim Neurons");
-        return parser_ok;
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemListNeurons(uint8_t displayIdx,
-                                                char *outKey, uint16_t outKeyLen,
-                                                char *outVal, uint16_t outValLen) {
-    if (displayIdx == 0) {
-        snprintf(outKey, outKeyLen, "Transaction type");
-        snprintf(outVal, outValLen, "List Own Neurons");
-        return parser_ok;
-    }
-
-    return parser_no_data;
-}
-
-static parser_error_t parser_getItemManageNeuron(const parser_context_t *ctx,
-                                                 uint8_t displayIdx,
-                                                 char *outKey, uint16_t outKeyLen,
-                                                 char *outVal, uint16_t outValLen,
-                                                 uint8_t pageIdx, uint8_t *pageCount) {
-    MEMZERO(outKey, outKeyLen);
-    MEMZERO(outVal, outValLen);
-    snprintf(outKey, outKeyLen, "?");
-    snprintf(outVal, outValLen, "?");
-    *pageCount = 1;
-
-    uint8_t numItems = 0;
-    CHECK_PARSER_ERR(parser_getNumItems(ctx, &numItems))
-    CHECK_APP_CANARY()
-
-    manageNeuron_e mn_type;
-    CHECK_PARSER_ERR(getManageNeuronType(&parser_tx_obj, &mn_type))
-
-    switch (mn_type) {
-        case Configure_IncreaseDissolveDelay:
-            return parser_getItemIncreaseNeuronTimer(displayIdx,
-                                                     outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-        case Configure_JoinCommunityFund :
-            return parser_getItemJoinCommunityFund(displayIdx,
-                                                   outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-        case Configure_StopDissolving :
-        case Configure_StartDissolving : {
-            return parser_getItemStartStopDissolve(displayIdx,
-                                                   outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-        }
-        case Spawn : {
-            return parser_getItemSpawn(displayIdx,
-                                       outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-        }
-        case Configure_RemoveHotKey:
-        case Configure_AddHotKey:
-            return parser_getItemAddRemoveHotkey(displayIdx, outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-
-        case Disburse :
-            return parser_getItemDisburse(displayIdx, outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-
-        case MergeMaturity :
-            return parser_getItemMergeMaturity(displayIdx, outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-
-        case RegisterVote :
-            return parser_getItemRegisterVote(displayIdx, outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-
-        case Follow:
-            return parser_getItemFollow(displayIdx, outKey, outKeyLen, outVal, outValLen, pageIdx, pageCount);
-
-        default:
-            return parser_no_data;
-    }
-}
-
 parser_error_t parser_getItem(const parser_context_t *ctx,
                               uint8_t displayIdx,
                               char *outKey, uint16_t outKeyLen,
@@ -1132,39 +208,14 @@ parser_error_t parser_getItem(const parser_context_t *ctx,
     switch (parser_tx_obj.txtype) {
         case call: {
             switch (parser_tx_obj.tx_fields.call.method_type) {
-                case pb_sendrequest : {
-                    const bool is_stake_tx = parser_tx_obj.special_transfer_type == neuron_stake_transaction;
-
-                    if (is_stake_tx) {
-                        return parser_getItemStakeNeuron(ctx, displayIdx,
-                                                         outKey, outKeyLen,
-                                                         outVal, outValLen,
-                                                         pageIdx, pageCount);
-                    }
-
-                    return parser_getItemTokenTransfer(ctx, displayIdx,
-                                                       outKey, outKeyLen,
-                                                       outVal, outValLen,
-                                                       pageIdx, pageCount);
-                }
-
-                case pb_manageneuron: {
-                    return parser_getItemManageNeuron(ctx, displayIdx,
-                                                      outKey, outKeyLen,
-                                                      outVal, outValLen,
-                                                      pageIdx, pageCount);
-                }
-
-                case pb_listneurons : {
-                    return parser_getItemListNeurons(displayIdx,
-                                                     outKey, outKeyLen,
-                                                     outVal, outValLen);
-                }
-
-                case pb_claimneurons : {
-                    return parser_getItemClaimNeuron(displayIdx,
-                                                     outKey, outKeyLen,
-                                                     outVal, outValLen);
+                case pb_sendrequest:
+                case pb_manageneuron:
+                case pb_listneurons:
+                case pb_claimneurons: {
+                    return parser_getItemProtobuf(displayIdx,
+                                                  outKey, outKeyLen,
+                                                  outVal, outValLen,
+                                                  pageIdx, pageCount);
                 }
 
                 case candid_manageneuron:
