+++ app/src/parser_print_helper.c
@@ -21,6 +21,9 @@
 #define CRC_LENGTH  4
 #define SUBACCOUNT_EXTRA_BYTE 0x7F
 
+// 365.25 * 24*60*60 = 31557600
+#define ICP_YEAR_IN_SECONDS ((uint64_t)31557600)
+
 parser_error_t print_u64(uint64_t value,
                         char *outVal, uint16_t outValLen,
                         uint8_t pageIdx, uint8_t *pageCount) {
@@ -44,13 +47,12 @@ parser_error_t print_ICP(uint64_t value,
     return parser_ok;
 }
 
-parser_error_t print_textual(uint8_t *data, uint16_t len,
+parser_error_t print_textual(const uint8_t *data, uint8_t len,
                              char *outVal, uint16_t outValLen,
                              uint8_t pageIdx, uint8_t *pageCount) {
     char tmpBuffer[100] = {0};
     uint16_t outLen = sizeof(tmpBuffer);
-    zxerr_t err = crypto_principalToTextual((const uint8_t *) data, len, (char *) tmpBuffer,
-                                            &outLen);
+    zxerr_t err = crypto_principalToTextual(data, len, (char *) tmpBuffer, &outLen);
     if (err != zxerr_ok) {
         return parser_unexpected_error;
     }
@@ -72,46 +74,7 @@ parser_error_t print_textual(uint8_t *data, uint16_t len,
     return parser_ok;
 }
 
-parser_error_t print_principal(uint8_t *data, uint16_t len,
-                             char *outVal, uint16_t outValLen,
-                             uint8_t pageIdx, uint8_t *pageCount) {
-
-    const uint8_t MAX_CHARS_PER_LINE = 38;
-    if (data == NULL || outVal == NULL || pageCount == NULL || outValLen < MAX_CHARS_PER_LINE) {
-        return parser_unexpected_error;
-    }
-    char tmpBuffer[100] = {0};
-    uint16_t outLen = sizeof(tmpBuffer);
-    zxerr_t err = crypto_principalToTextual((const uint8_t *) data, len, (char *) tmpBuffer,
-                                            &outLen);
-    if (err != zxerr_ok) {
-        return parser_unexpected_error;
-    }
-
-    char buffer[100] = {0};
-    err = addr_to_textual(buffer, sizeof(buffer), (const char *) tmpBuffer, outLen);
-    if (err != zxerr_ok) {
-        return parser_unexpected_error;
-    }
-
-    // Alignment logic
-    if (buffer[17] == '-') buffer[17] = ' ';
-    if (buffer[53] == '-') buffer[53] = ' ';
-    // Remove one buffer[35] ('-' character)
-    MEMMOVE(buffer+35, buffer+36, sizeof(buffer) - 36);
-
-    inplace_insert_char(buffer, sizeof(buffer), 18, ':');
-    inplace_insert_char(buffer, sizeof(buffer), 19, ' ');
-
-    inplace_insert_char(buffer, sizeof(buffer), 55, ':');
-    inplace_insert_char(buffer, sizeof(buffer), 56, ' ');
-
-    pageString(outVal, MAX_CHARS_PER_LINE, buffer, pageIdx, pageCount);
-
-    return parser_ok;
-}
-
-parser_error_t print_canisterId(const uint8_t *data, uint16_t len,
+parser_error_t print_canisterId(const uint8_t *data, uint8_t len,
                              char *outVal, uint16_t outValLen,
                              uint8_t pageIdx, uint8_t *pageCount) {
     char tmpBuffer[100] = {0};
@@ -142,7 +105,9 @@ parser_error_t print_canisterId(const uint8_t *data, uint16_t len,
 
 zxerr_t print_hexstring(char *out, uint16_t outLen, const uint8_t *data, uint16_t dataLen) {
     MEMZERO(out, outLen);
-    const uint32_t writtenBytes = array_to_hexstr(out, outLen, data, dataLen);
+    if (dataLen > 255) return zxerr_out_of_bounds;
+
+    const uint32_t writtenBytes = array_to_hexstr(out, outLen, data, (uint8_t) dataLen);
     if (writtenBytes != dataLen * 2) {
         return zxerr_out_of_bounds;
     }
@@ -187,8 +152,8 @@ parser_error_t subaccount_hexstring(const uint8_t *subaccount, const uint16_t su
     // 3 chunks per page and do not add delimiter on the last one from each page
     uint8_t delimiterCount = 0;
     for (uint16_t i = 0; i < subaccountLen; i+= 3) {
-        const uint8_t bytesToProcess = (subaccountLen - i) >= 3 ? 3 : (subaccountLen - i);
-        for (uint8_t j = 0; j < bytesToProcess; j++) {
+        const uint16_t bytesToProcess = (subaccountLen - i) >= 3 ? 3 : (subaccountLen - i);
+        for (uint16_t j = 0; j < bytesToProcess; j++) {
             snprintf((char*) output, 3, "%02x", *subaccount);
             subaccount++;
             output += 2;
@@ -216,7 +181,8 @@ static parser_error_t page_with_delimiters(char *input, const uint16_t inputLen,
         return parser_unexpected_buffer_end;
     }
 
-    *pageCount = inputLen / CHARS_PER_PAGE + (inputLen % CHARS_PER_PAGE ? 1 : 0);
+    const size_t inputStrLen = strnlen(input, inputLen);
+    *pageCount = (uint8_t) (inputStrLen / CHARS_PER_PAGE) + ((inputStrLen % CHARS_PER_PAGE) ? 1 : 0);
     if (pageIdx >= *pageCount) {
         return parser_display_idx_out_of_range;
     }
@@ -228,7 +194,8 @@ static parser_error_t page_with_delimiters(char *input, const uint16_t inputLen,
             output += 1;
         }
 
-        const bool endOfInput = strlen(input) < 6;
+        const size_t remainingChars = inputStrLen - (pageIdx * CHARS_PER_PAGE) - (idx * CHARS_PER_CHUNK);
+        const bool endOfInput = remainingChars < 6;
         const bool skipDash = (idx == 2 || idx == 5);
 
         if (skipDash || endOfInput) {
@@ -255,12 +222,12 @@ parser_error_t print_principal_with_subaccount(const uint8_t *sender, uint16_t s
         return parser_unexpected_error;
     }
 
-    //[ CRC | sender | shrink(fromSubaccount) | bytes(shrink(fromSubaccount)) | EXTRA_BYTE(0x7F) ]
+    // [ CRC | sender | shrink(fromSubaccount) | bytes(shrink(fromSubaccount)) | EXTRA_BYTE(0x7F) ]
     uint8_t tmpArray[CRC_LENGTH + DFINITY_PRINCIPAL_LEN + DFINITY_SUBACCOUNT_LEN + 2] = {0};
 
     MEMCPY(tmpArray + CRC_LENGTH, sender, senderLen);
 
-    int8_t shrinkBytes = 0;
+    int shrinkBytes = 0;
     if (fromSubaccount != NULL) {
         for (shrinkBytes = fromSubaccountLen - 1; shrinkBytes >= 0; shrinkBytes--) {
             if (*(fromSubaccount + shrinkBytes)) {
@@ -270,21 +237,93 @@ parser_error_t print_principal_with_subaccount(const uint8_t *sender, uint16_t s
         shrinkBytes++;
 
         if (shrinkBytes > 0) {
+            if (shrinkBytes > DFINITY_SUBACCOUNT_LEN) return parser_unexpected_value;
             MEMCPY(tmpArray + CRC_LENGTH + DFINITY_PRINCIPAL_LEN, fromSubaccount, shrinkBytes);
             // Add fromSubaccount length after shrinked and EXTRA_BYTE
-            tmpArray[CRC_LENGTH + DFINITY_PRINCIPAL_LEN + shrinkBytes] = shrinkBytes;
+            tmpArray[CRC_LENGTH + DFINITY_PRINCIPAL_LEN + shrinkBytes] = (uint8_t) shrinkBytes;
             tmpArray[CRC_LENGTH + DFINITY_PRINCIPAL_LEN + shrinkBytes + 1] = SUBACCOUNT_EXTRA_BYTE;
         }
     }
 
 
     const uint8_t tmpArrayLen = shrinkBytes
-                                ? (CRC_LENGTH + DFINITY_PRINCIPAL_LEN + shrinkBytes + 2)
+                                ? (CRC_LENGTH + DFINITY_PRINCIPAL_LEN + (uint8_t) shrinkBytes + 2)
                                 : (CRC_LENGTH + DFINITY_PRINCIPAL_LEN);
 
     char buffer[120] = {0};
     uint16_t bufferSize = sizeof(buffer);
     crypto_toTextual(tmpArray, tmpArrayLen, buffer, &bufferSize);
 
-    return page_with_delimiters(buffer, bufferSize, outVal, outValLen, pageIdx, pageCount);
+    return page_with_delimiters(buffer, sizeof(buffer), outVal, outValLen, pageIdx, pageCount);
+}
+
+parser_error_t parser_printDelay(uint64_t value, char *buffer, uint16_t bufferSize) {
+    MEMZERO(buffer, bufferSize);
+    uint16_t index = 0;
+    uint64_t years = value / ICP_YEAR_IN_SECONDS;
+    if (years >= 1) {
+        index += fpuint64_to_str(buffer, bufferSize, years, 0);
+        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
+        MEMCPY(buffer + index, (char *) "y", 1);
+        index += 1;
+    }
+    value %= ICP_YEAR_IN_SECONDS;
+
+    uint64_t days = value / (uint64_t) (60 * 60 * 24);
+    if (days > 0) {
+        if (index > 0) {
+            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
+            MEMCPY(buffer + index, (char *) ", ", 2);
+            index += 2;
+        }
+        index += fpuint64_to_str(buffer + index, bufferSize - index, days, 0);
+        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
+        MEMCPY(buffer + index, (char *) "d", 1);
+        index += 1;
+    }
+    value %= (uint64_t) (60 * 60 * 24);
+
+    uint64_t hours = value / (uint64_t) (60 * 60);
+    if (hours > 0) {
+        if (index > 0) {
+            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
+            MEMCPY(buffer + index, (char *) ", ", 2);
+            index += 2;
+        }
+        index += fpuint64_to_str(buffer + index, bufferSize - index, hours, 0);
+        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
+        MEMCPY(buffer + index, (char *) "h", 1);
+        index += 1;
+    }
+    value %= (uint64_t) (60 * 60);
+
+    uint64_t minutes = value / (uint64_t) (60);
+    if (minutes > 0) {
+        if (index > 0) {
+            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
+            MEMCPY(buffer + index, (char *) ", ", 2);
+            index += 2;
+        }
+        index += fpuint64_to_str(buffer + index, bufferSize - index, minutes, 0);
+        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
+        MEMCPY(buffer + index, (char *) "m", 1);
+        index += 1;
+    }
+    value %= (uint64_t) (60);
+
+    uint64_t seconds = value;
+    if (seconds > 0) {
+        if (index > 0) {
+            PARSER_ASSERT_OR_ERROR(index + 2 < bufferSize, parser_unexpected_buffer_end)
+            MEMCPY(buffer + index, (char *) ", ", 2);
+            index += 2;
+        }
+        index += fpuint64_to_str(buffer + index, bufferSize - index, seconds, 0);
+        PARSER_ASSERT_OR_ERROR(index + 1 < bufferSize, parser_unexpected_buffer_end)
+        MEMCPY(buffer + index, (char *) "s", 1);
+        index += 1;
+    }
+
+    buffer[index] = 0;
+    return parser_ok;
 }
