+++ app/src/parser_print_helper.h
@@ -27,17 +27,14 @@ extern "C" {
 
 parser_error_t print_u64(uint64_t value, char *outVal, uint16_t outValLen, uint8_t pageIdx, uint8_t *pageCount);
 parser_error_t print_ICP(uint64_t value, char *outVal, uint16_t outValLen, uint8_t pageIdx, uint8_t *pageCount);
-parser_error_t print_textual(uint8_t *data, uint16_t len, char *outVal, uint16_t outValLen, uint8_t pageIdx, uint8_t *pageCount);
+parser_error_t print_textual(const uint8_t *data, uint8_t len, char *outVal, uint16_t outValLen, uint8_t pageIdx, uint8_t *pageCount);
 parser_error_t parser_printDelay(uint64_t value, char *buffer, uint16_t bufferSize);
 zxerr_t print_hexstring(char *out, uint16_t outLen, const uint8_t *data, uint16_t dataLen);
 
 parser_error_t subaccount_hexstring(const uint8_t *subaccount, const uint16_t subaccountLen,
                                     uint8_t *output, const uint16_t outputLen, uint8_t *pageCount);
 
-parser_error_t print_principal(uint8_t *data, uint16_t len, char *outVal, uint16_t outValLen,
-                               uint8_t pageIdx, uint8_t *pageCount);
-
-parser_error_t print_canisterId(const uint8_t *data, uint16_t len, char *outVal, uint16_t outValLen,
+parser_error_t print_canisterId(const uint8_t *data, uint8_t len, char *outVal, uint16_t outValLen,
                                 uint8_t pageIdx, uint8_t *pageCount);
 
 parser_error_t print_principal_with_subaccount(const uint8_t *sender, uint16_t senderLen,
